# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Leaflow Amber",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/openai-compatible/v1/chat/completions": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "兼容 OpenAI Chat Completion 接口，认证需要使用 Assistant Share Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "OpenAI Chat Completion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "指定聊天中的用户 IP 地址，不指定则自动获取。此 IP 地址只会增加至 Prompt 中，如果不希望增加，请关闭系统自带 Prompt 选项",
                        "name": "X-User-IP",
                        "in": "header"
                    },
                    {
                        "description": "Chat",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.OpenAIChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.OpenAIChatCompletionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.ResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schema.ResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.OpenAIChatCompletionRequest": {
            "type": "object",
            "properties": {
                "max_tokens": {
                    "description": "Optional",
                    "type": "integer"
                },
                "messages": {
                    "description": "Required",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.OpenAIChatCompletionRequestMessage"
                    }
                },
                "model": {
                    "description": "Required",
                    "type": "string"
                },
                "n": {
                    "description": "Optional",
                    "type": "integer"
                },
                "stream": {
                    "description": "Optional",
                    "type": "boolean"
                },
                "temperature": {
                    "description": "Optional",
                    "type": "number"
                },
                "top_p": {
                    "description": "Optional",
                    "type": "number"
                }
            }
        },
        "schema.OpenAIChatCompletionRequestMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Required"
                },
                "image_url": {
                    "$ref": "#/definitions/schema.OpenAIImageUrl"
                },
                "role": {
                    "description": "Required",
                    "type": "string"
                }
            }
        },
        "schema.OpenAIChatCompletionResponse": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.OpenAIChatCompletionResponseChoice"
                    }
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "usage": {
                    "$ref": "#/definitions/schema.TokenUsage"
                }
            }
        },
        "schema.OpenAIChatCompletionResponseChoice": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer"
                },
                "message": {
                    "$ref": "#/definitions/schema.OpenAIChatCompletionRequestMessage"
                }
            }
        },
        "schema.OpenAIImageUrl": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "schema.ResponseBody": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "schema.TokenUsage": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer"
                },
                "prompt_tokens": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}